+++ readium/shared/src/main/java/org/readium/r2/shared/publication/Metadata.kt
@@ -29,6 +29,9 @@ import org.readium.r2.shared.util.logging.log
 /**
  * https://readium.org/webpub-manifest/schema/metadata.schema.json
  *
+ * @param tdm Publications can indicate whether they allow third parties to use their content for
+ * text and data mining purposes using the [TDM Rep protocol](https://www.w3.org/community/tdmrep/),
+ * as defined in a [W3C Community Group Report](https://www.w3.org/community/reports/tdmrep/CG-FINAL-tdmrep-20240510/).
  * @param otherMetadata Additional metadata for extensions, as a JSON dictionary.
  */
 @Parcelize
@@ -62,6 +65,7 @@ public data class Metadata(
     val duration: Double? = null,
     val numberOfPages: Int? = null,
     val belongsTo: Map<String, List<Collection>> = emptyMap(),
+    val tdm: Tdm? = null,
     val otherMetadata: @WriteWith<JSONParceler> Map<String, Any> = mapOf(),
 ) : JSONable, Parcelable {
 
@@ -97,6 +101,7 @@ public data class Metadata(
         belongsTo: Map<String, List<Collection>> = emptyMap(),
         belongsToCollections: List<Collection> = emptyList(),
         belongsToSeries: List<Collection> = emptyList(),
+        tdm: Tdm? = null,
         otherMetadata: Map<String, Any> = mapOf(),
     ) : this(
         identifier = identifier,
@@ -138,6 +143,7 @@ public data class Metadata(
                 }
             }
             .toMap(),
+        tdm = tdm,
         otherMetadata = otherMetadata
     )
 
@@ -198,6 +204,7 @@ public data class Metadata(
         put("duration", duration)
         put("numberOfPages", numberOfPages)
         putIfNotEmpty("belongsTo", belongsTo)
+        putIfNotEmpty("tdm", tdm)
     }
 
     /**
@@ -298,6 +305,7 @@ public data class Metadata(
             val duration = json.optPositiveDouble("duration", remove = true)
             val numberOfPages = json.optPositiveInt("numberOfPages", remove = true)
 
+            val tdm = Tdm.fromJSON(json.remove("tdm"), warnings)
             val belongsToJson = (
                 json.remove("belongsTo") as? JSONObject
                     ?: json.remove("belongs_to") as? JSONObject
@@ -345,6 +353,7 @@ public data class Metadata(
                 duration = duration,
                 numberOfPages = numberOfPages,
                 belongsTo = belongsTo.toMap(),
+                tdm = tdm,
                 otherMetadata = json.toMap()
             )
         }